# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Severinno Marketplace API",
    "version": "0.2.0",
    "description": "API do Marketplace de serviços com geolocalização"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "API Server"
    }
  ],
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Registrar novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado"
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado"
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
...
    "/api/providers": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Listar prestadores",
        "parameters": [
          {
            "name": "lat",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "lng",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "radius",
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "rating",
                "distance"
              ]
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de prestadores"
          }
        }
      }
    },
    "/api/categories": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Listar categorias",
        "responses": {
          "200": {
            "description": "Árvore de categorias"
          }
        }
      },
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Categories"
        ]
      }
    },
...
    "/api/auth/forgot-password": {
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/newsletter": {
      "post": {
        "summary": "POST /api/newsletter",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Newsletter"
        ]
      }
    },
...
    "/api/sentry": {
      "post": {
        "summary": "Tunnel for Sentry/GlitchTip client-side errors.",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Sentry"
        ]
      }
    },
...
    "/api/services": {
      "get": {
        "summary": "Retrieve",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "tags": [
          "Services"
        ]
      },
      "post": {
        "summary": "Create",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Services"
        ]
      }
    },
...
    "/api/webhooks/evolution": {
      "post": {
        "summary": "Webhook Evolution API",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/webhooks/lytex": {
      "post": {
        "summary": "Webhook Lytex Pagamentos",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    },
    "/api/webhooks/sentry-alert": {
      "post": {
        "summary": "POST /api/webhooks/sentry-alert — Recebe alertas do GlitchTip/Sentry e encaminha",
        "responses": {
          "200": {
            "description": "Success"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "tags": [
          "Webhooks"
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "LoginInput": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "E-mail do usuário"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "description": "Senha"
          }
        }
      },
      "RegisterInput": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "confirmPassword",
          "role"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 2
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6
          },
          "confirmPassword": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "CLIENT",
              "PROVIDER"
            ]
          },
          "cpfCnpj": {
            "type": "string"
          },
          "whatsapp": {
            "type": "string"
          },
          "cep": {
            "type": "string"
          },
          "street": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string",
            "maxLength": 2
          },
          "bio": {
            "type": "string",
            "maxLength": 600
          },
          "radiusKm": {
            "type": "number"
          }
        }
      },
...
    }
  }
}